controller/pyproject.toml
[tool.poetry]
name = "controller"
version = "0.1.0"
description = ""
authors = ["Ameya Rao <ameyarao98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^22.6.2"
Sanic-Cors = "^2.1.0"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
